refactor(spells): Extracted the ImportPage component into its own file
